+++ apps/server/bin/server.dart
@@ -1,18 +1,36 @@
+import 'dart:async';
 import 'dart:developer';
 import 'dart:io';
 
 import 'package:logging/logging.dart';
 import 'package:mongo_dart/mongo_dart.dart';
 
-import 'package:protos_weebi/grpc.dart' show Server;
+import 'package:protos_weebi/grpc.dart' show GrpcError, Server, ServiceCall, ServiceMethod;
 import 'package:article_service/article_service.dart';
 import 'package:contact_service/contact_service.dart';
+import 'package:server/server_interceptors.dart';
 import 'package:ticket_service/ticket_service.dart';
 import 'package:fence_service/fence_service.dart';
-// import 'package:fence_service/user_testing.dart';
+// import 'package:fence_service/mongo_local_testing.dart';
 import 'package:fence_service/weebi_app_service.dart';
 
 import '../constants/app_environment.dart';
+
+FutureOr<GrpcError?> corsInterceptor(ServiceCall call, ServiceMethod method) {
+  call.headers!.addAll({
+    HttpHeaders.accessControlAllowOriginHeader: '*',
+    HttpHeaders.accessControlAllowMethodsHeader: 'GET, POST, PUT, DELETE, OPTIONS',
+    HttpHeaders.accessControlAllowHeadersHeader: '*',
+    HttpHeaders.accessControlExposeHeadersHeader: '*',
+    HttpHeaders.accessControlAllowCredentialsHeader: 'true',
+  });
+
+  if (call.clientMetadata?[':method'] == 'OPTIONS') {
+    return null;
+  }
+
+  return null;
+}
 
 void main(List<String> arguments) async {
   Logger.root.level = Level.ALL;
@@ -37,6 +55,7 @@ void main(List<String> arguments) async {
   final interceptors = [
     loggingInterceptor,
     authInterceptor,
+    corsInterceptor
   ];
 
   final articleService = ArticleService(db);