+++ apps/server/bin/server_offline.dart
@@ -6,9 +6,10 @@ import 'package:logging/logging.dart';
 import 'package:protos_weebi/grpc.dart' show Server;
 import 'package:article_service/article_service.dart';
 import 'package:contact_service/contact_service.dart';
+import 'package:server/server_interceptors.dart';
 import 'package:ticket_service/ticket_service.dart';
 import 'package:fence_service/fence_service.dart';
-import 'package:fence_service/user_testing.dart';
+import 'package:fence_service/mongo_local_testing.dart';
 import 'package:fence_service/weebi_app_service.dart';
 
 void main(List<String> arguments) async {