+++ apps/server/lib/client_interceptor_example.dart
@@ -1,4 +1,4 @@
-library client_interceptors;
+library server_interceptors;
 
 export 'src/client_interceptor/auth_interceptor.dart';
 export 'src/client_interceptor/log_interceptor.dart';