+++ apps/server/utils/create_user.dart
@@ -1,7 +1,8 @@
 import 'package:mongo_dart/mongo_dart.dart';
 
 import 'package:fence_service/fence_service.dart';
-import 'package:fence_service/user_testing.dart';
+import 'package:fence_service/mongo_local_testing.dart';
+import 'package:protos_weebi/data_dummy.dart';
 
 void main() async {
   final db = TestHelper.localDb;
@@ -9,12 +10,13 @@ void main() async {
   late FenceService fenceService;
   await db.open();
   final isConnected = await connection.connect();
-  print(isConnected);
+  print('isConnected $isConnected');
   fenceService = FenceService(db);
   //await db.createCollection(fenceService.collection.collectionName);
 
-  await fenceService.userCollection
+  final dd = await fenceService.userCollection
       .insertOne(Dummy.userPublic.toProto3Json() as Map<String, dynamic>);
+  print(dd.writeError);
   await fenceService.userCollection.update(
     where.eq('userId', Dummy.userPublic.userId),
     ModifierBuilder().set('password', Dummy.password),