+++ packages/fence_service/lib/fence_service.dart
@@ -1,11 +1,7 @@
 library fence_service;
 
 export 'src/fence_service_base.dart';
-export 'src/log_interceptor.dart';
-export 'src/auth_interceptor.dart';
-export 'src/data/roles_default.dart';
 export 'src/jwt.dart';
 export 'src/bearer_ext.dart';
 export 'src/user_perm_ext.dart';
 export 'src/extension.dart';
-export 'src/password_encrypter.dart';