+++ packages/fence_service/lib/mongo_local_testing.dart
@@ -1,0 +1,3 @@
+library mongo_local_testing;
+
+export 'src/testing/test_helper.dart';