+++ packages/protos/protos_weebi/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.1
+
+- complete device enrolment
+- add client interceptors
+
 ## 1.0.0
 
 - Initial version